+++ platform/distutils.ts
@@ -1,4 +1,5 @@
 import {fs_exists, fs_stat} from "./io";
+import {printError} from "./print";
 
 export class log {
   public static info(msg: string){
@@ -10,7 +11,9 @@ export class dep_util {
   public static newer(input_path: string, output_path: string){
     const exists_input_file = fs_exists(input_path);
     if(!exists_input_file){
-      throw new Error("source does not exist");
+      const errMsg = "source does not exist";
+      printError(`Error: ${errMsg}`);
+      throw new Error(errMsg);
     }
     
     const exists_output_file = fs_exists(output_path);