+++ stages/stage_0.ts
@@ -11,6 +11,7 @@ import {
   Bytes,
 } from "clvm";
 import * as binutils from "../clvm_tools/binutils";
+import {printError} from "../platform/print";
 
 export const run = binutils.assemble("(a 2 3)");
 export const brun = run;
@@ -34,7 +35,9 @@ export function run_program(
   
   if(strict){
     const fatal_error = (op: Bytes, args: SExp) => {
-      throw new EvalError("unimplemented operator", SExp.to(op));
+      const errMsg = "unimplemented operator";
+      printError(`EvalError: ${errMsg} ${op}`);
+      throw new EvalError(errMsg, SExp.to(op));
     }
     operator_lookup = OperatorDict(operator_lookup, {unknown_op_handler: fatal_error});
   }